Run Sonobuoy with Calico instead of flannel
This commit switches networking for Sonobuoy from flannel to Calico.
This is because the flannel CNI plugin has been moved to a separate
project from CNI Plugins since v1.0.0.

This commit does not delete "example/kube-flannel.yaml" because
it is used by the recipe of "run-on-vagrant" in sonobuoy/Makefile.
This target is not maintained now and will be deleted soon.
The manifest file will be deleted at the same time.

Signed-off-by: morimoto-cybozu <kenji_morimoto@cybozu.co.jp>